Made Services and the first service - a `QuestionFaclory` class - made `Services` folder in the project - made QuestionFactory.swift file in this folder - made `QuestionFactory` class - moved questions array with mock data into new class - made new method  requestNextQuestion()
